# [ hgemm ] Use hgemm kernel in transpose cases