- With SIMD version of fp16 transpose code, using hgemm kernel in transpose case would be more useful.
- Note that we should develop a data packing code for this case for further optimization.
- Still, this gives huge difference btw

### hgemm_transA

| dim | before | after |
| --- | --- | --- |
| 1024 square | 192 ms | **35 ms** |
| 768 square | 71 ms | **13.0 ms** |
| 256,1440,256 | 15 ms | **2.8 ms** |
| 256,256,1440 | 15 ms | **3.0 ms** |
| 1920,1560,88 | 46 ms | **9.8 ms** |

### hgemm_transB

| dim | before | after |
| --- | --- | --- |
| 1024 square | 88 ~ 113 ms | **30 ms** |
| 768 square | 31 ms | **12 ms** |
| 256,1440,256 | 6.4 ms | **2.6 ms** 
| 256,256,1440 | 7.1 ms | **3.0 ms** |
| 1560,2048,512 | 103 ~ 188 ms | **45 ms** |

### hgemm_transAB

| dim | before | after |
| --- | --- | --- |
| 1024 square | 2599 ms | **37 ms** |
| 768 square | 814 ms | **12 ms** |
| 256,1440,256 | 160 ms | **2.8 ms** |
| 256,256,1440 | 70 ms | **3.0 ms** |

**Self evaluation:**
1. Build test:     [X]Passed [ ]Failed [ ]Skipped
2. Run test:     [X]Passed [ ]Failed [ ]Skipped